src/application/requirements.txt
asyncio
grpcio
grpcio-tools
pydantic
pydantic-settings
sounddevice